+++ src/app/app.routes.ts
@@ -7,7 +7,7 @@ export const routes: Routes = [
   },
   {
     path: 'trades',
-    loadComponent: () => import('./orders/orders-list/orders-list').then(m => m.OrdersList)
+    loadComponent: () => import('./trades/trades-list/trades-list').then(m => m.TradesList)
   },
   {
     path: 'trades/new',
@@ -15,6 +15,6 @@ export const routes: Routes = [
   },
   {
     path: 'trades/:id',
-    loadComponent: () => import('./orders/order-detail/order-detail').then(m => m.OrderDetail)
+    loadComponent: () => import('./trades/trade-detail/trade-detail').then(m => m.TradeDetail)
   }
 ];