+++ src/app/trades/trade-detail/trade-detail.ts
@@ -2,20 +2,20 @@ import { Component, signal } from '@angular/core';
 import { ActivatedRoute, Router } from '@angular/router';
 
 @Component({
-  selector: 'app-order-detail',
+  selector: 'app-trade-detail',
   imports: [],
-  templateUrl: './order-detail.html',
-  styleUrl: './order-detail.scss'
+  templateUrl: './trade-detail.html',
+  styleUrl: './trade-detail.scss'
 })
-export class OrderDetail {
-  protected readonly orderId = signal<string>('');
+export class TradeDetail {
+  protected readonly tradeId = signal<string>('');
 
   constructor(
     private readonly route: ActivatedRoute,
     private readonly router: Router
   ) {
     const id = this.route.snapshot.paramMap.get('id') ?? '';
-    this.orderId.set(id);
+    this.tradeId.set(id);
   }
 
   goBack(): void {